refactor: move `CookieJar` in its own file
