+++ lib/requests.dart
@@ -3,3 +3,4 @@ library requests;
 export 'src/requests.dart';
 export 'src/response.dart';
 export 'src/cookie.dart';
+export 'src/cookie_jar.dart';
